Address review feedback
- GetExports no-constraint fast path: return the existing ImmutableList directly (it is
  immutable and the method returns IReadOnlyList<>), and pre-size the filtered list to the
  per-contract candidate count.
- Move the benchmark project from perf/ to test/ (turns off packing automatically and applies
  the test/.editorconfig analyzer relaxations); drop the IsPackable/EnablePackageValidation/
  analyzer/code-style/doc/NoWarn and GC-mode overrides so it inherits repo conventions.
- Target net10.0; split the mixed ItemGroup into one item type per group.
- Program.cs top-level statement; file-scoped namespaces throughout (Parts.cs keeps its
  compact generated layout via file-level pragma suppressions).
- Dispose the ExportProvider in the Runtime benchmark; add GetExportsUnconstrained/
  GetExportsConstrained benchmarks to guard the no-constraint fast path.
- Remove the CentralPackageTransitivePinningEnabled overrides: on net10 the framework supplies
  the higher System.Collections.Immutable/Reflection.Metadata versions (package pruning), so
  central pinning no longer triggers a downgrade.

Builds clean on net10.0 with analyzers/code-style/docs enabled; all 2057 Composition tests pass.

Copilot-Session: 9d04317c-fb22-4281-be23-493d4da40a66
