+++ Directory.Packages.props
@@ -11,10 +11,6 @@
     <CodefixTestingVersion>1.1.4</CodefixTestingVersion>
   </PropertyGroup>
 
-  <!-- BenchmarkDotNet's generated project inherits CPM; let it resolve its own transitive versions to avoid a downgrade. -->
-  <PropertyGroup Condition="'$(MSBuildProjectName)' == 'BenchmarkDotNet.Autogenerated'">
-    <CentralPackageTransitivePinningEnabled>false</CentralPackageTransitivePinningEnabled>
-  </PropertyGroup>
   <ItemGroup>
     <PackageVersion Include="DiffPlex" Version="1.9.0" />
     <PackageVersion Include="BenchmarkDotNet" Version="0.15.8" />
