--- perf/Microsoft.VisualStudio.Composition.Benchmarks/Microsoft.VisualStudio.Composition.Benchmarks.csproj
@@ -1,31 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>net8.0</TargetFramework>
-    <Nullable>disable</Nullable>
-    <ImplicitUsings>disable</ImplicitUsings>
-    <GenerateDocumentationFile>false</GenerateDocumentationFile>
-    <EnforceCodeStyleInBuild>false</EnforceCodeStyleInBuild>
-    <EnableNETAnalyzers>false</EnableNETAnalyzers>
-    <RunAnalyzers>false</RunAnalyzers>
-    <IsPackable>false</IsPackable>
-    <EnablePackageValidation>false</EnablePackageValidation>
-    <ProduceReferenceAssembly>false</ProduceReferenceAssembly>
-    <!-- Let BenchmarkDotNet pick its own transitive versions; the repo's central pins force a downgrade. -->
-    <CentralPackageTransitivePinningEnabled>false</CentralPackageTransitivePinningEnabled>
-    <NoWarn>$(NoWarn);CS1591;SA0001;CS8032;NU1701;MSB3277</NoWarn>
-    <Optimize>true</Optimize>
-    <ServerGarbageCollection>false</ServerGarbageCollection>
-    <ConcurrentGarbageCollection>false</ConcurrentGarbageCollection>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\..\src\Microsoft.VisualStudio.Composition\Microsoft.VisualStudio.Composition.csproj" />
-    <PackageReference Include="System.Composition.AttributedModel" />
-    <PackageReference Include="System.ComponentModel.Composition" />
-    <PackageReference Include="BenchmarkDotNet" />
-    <PackageReference Include="Microsoft.VisualStudio.DiagnosticsHub.BenchmarkDotNetDiagnosers" />
-  </ItemGroup>
-
-</Project>
