--- perf/Microsoft.VisualStudio.Composition.Benchmarks/Parts.cs
@@ -1,344 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT license. See LICENSE file in the project root for full license information.
-
-// Synthetic MEF catalog (~312 parts) exercised by the benchmarks.
-namespace Microsoft.VisualStudio.Composition.Benchmarks
-{
-    using System.Collections.Generic;
-    using MefV1 = System.ComponentModel.Composition;
-    using MefV2 = System.Composition;
-
-    public interface IService { string Name { get; } }
-    public interface ILogger { void Log(string message); }
-    public interface ICache { object Get(string key); }
-    public interface IProcessor { string Process(string input); }
-    public interface IServiceMetadata { int Order { get; } string Category { get; } }
-
-    [MefV2.Export(typeof(ILogger)), MefV2.Shared] public class ConsoleLogger : ILogger { public void Log(string m) { } }
-    [MefV2.Export(typeof(ICache)), MefV2.Shared] public class MemoryCache : ICache { public object Get(string k) => null; }
-
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 1), MefV2.ExportMetadata("Category", "beta"), MefV2.Shared] public class Service1 : IService { public string Name => "1"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 2), MefV2.ExportMetadata("Category", "gamma"), MefV2.Shared] public class Service2 : IService { public string Name => "2"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 3), MefV2.ExportMetadata("Category", "delta"), MefV2.Shared] public class Service3 : IService { public string Name => "3"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 4), MefV2.ExportMetadata("Category", "alpha"), MefV2.Shared] public class Service4 : IService { public string Name => "4"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 5), MefV2.ExportMetadata("Category", "beta"), MefV2.Shared] public class Service5 : IService { public string Name => "5"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 6), MefV2.ExportMetadata("Category", "gamma"), MefV2.Shared] public class Service6 : IService { public string Name => "6"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 7), MefV2.ExportMetadata("Category", "delta"), MefV2.Shared] public class Service7 : IService { public string Name => "7"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 8), MefV2.ExportMetadata("Category", "alpha"), MefV2.Shared] public class Service8 : IService { public string Name => "8"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 9), MefV2.ExportMetadata("Category", "beta"), MefV2.Shared] public class Service9 : IService { public string Name => "9"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 10), MefV2.ExportMetadata("Category", "gamma"), MefV2.Shared] public class Service10 : IService { public string Name => "10"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 11), MefV2.ExportMetadata("Category", "delta"), MefV2.Shared] public class Service11 : IService { public string Name => "11"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 12), MefV2.ExportMetadata("Category", "alpha"), MefV2.Shared] public class Service12 : IService { public string Name => "12"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 13), MefV2.ExportMetadata("Category", "beta"), MefV2.Shared] public class Service13 : IService { public string Name => "13"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 14), MefV2.ExportMetadata("Category", "gamma"), MefV2.Shared] public class Service14 : IService { public string Name => "14"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 15), MefV2.ExportMetadata("Category", "delta"), MefV2.Shared] public class Service15 : IService { public string Name => "15"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 16), MefV2.ExportMetadata("Category", "alpha"), MefV2.Shared] public class Service16 : IService { public string Name => "16"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 17), MefV2.ExportMetadata("Category", "beta"), MefV2.Shared] public class Service17 : IService { public string Name => "17"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 18), MefV2.ExportMetadata("Category", "gamma"), MefV2.Shared] public class Service18 : IService { public string Name => "18"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 19), MefV2.ExportMetadata("Category", "delta"), MefV2.Shared] public class Service19 : IService { public string Name => "19"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 20), MefV2.ExportMetadata("Category", "alpha"), MefV2.Shared] public class Service20 : IService { public string Name => "20"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 21), MefV2.ExportMetadata("Category", "beta"), MefV2.Shared] public class Service21 : IService { public string Name => "21"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 22), MefV2.ExportMetadata("Category", "gamma"), MefV2.Shared] public class Service22 : IService { public string Name => "22"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 23), MefV2.ExportMetadata("Category", "delta"), MefV2.Shared] public class Service23 : IService { public string Name => "23"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 24), MefV2.ExportMetadata("Category", "alpha"), MefV2.Shared] public class Service24 : IService { public string Name => "24"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 25), MefV2.ExportMetadata("Category", "beta"), MefV2.Shared] public class Service25 : IService { public string Name => "25"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 26), MefV2.ExportMetadata("Category", "gamma"), MefV2.Shared] public class Service26 : IService { public string Name => "26"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 27), MefV2.ExportMetadata("Category", "delta"), MefV2.Shared] public class Service27 : IService { public string Name => "27"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 28), MefV2.ExportMetadata("Category", "alpha"), MefV2.Shared] public class Service28 : IService { public string Name => "28"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 29), MefV2.ExportMetadata("Category", "beta"), MefV2.Shared] public class Service29 : IService { public string Name => "29"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 30), MefV2.ExportMetadata("Category", "gamma"), MefV2.Shared] public class Service30 : IService { public string Name => "30"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 31), MefV2.ExportMetadata("Category", "delta"), MefV2.Shared] public class Service31 : IService { public string Name => "31"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 32), MefV2.ExportMetadata("Category", "alpha"), MefV2.Shared] public class Service32 : IService { public string Name => "32"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 33), MefV2.ExportMetadata("Category", "beta"), MefV2.Shared] public class Service33 : IService { public string Name => "33"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 34), MefV2.ExportMetadata("Category", "gamma"), MefV2.Shared] public class Service34 : IService { public string Name => "34"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 35), MefV2.ExportMetadata("Category", "delta"), MefV2.Shared] public class Service35 : IService { public string Name => "35"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 36), MefV2.ExportMetadata("Category", "alpha"), MefV2.Shared] public class Service36 : IService { public string Name => "36"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 37), MefV2.ExportMetadata("Category", "beta"), MefV2.Shared] public class Service37 : IService { public string Name => "37"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 38), MefV2.ExportMetadata("Category", "gamma"), MefV2.Shared] public class Service38 : IService { public string Name => "38"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 39), MefV2.ExportMetadata("Category", "delta"), MefV2.Shared] public class Service39 : IService { public string Name => "39"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 40), MefV2.ExportMetadata("Category", "alpha"), MefV2.Shared] public class Service40 : IService { public string Name => "40"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 41), MefV2.ExportMetadata("Category", "beta"), MefV2.Shared] public class Service41 : IService { public string Name => "41"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 42), MefV2.ExportMetadata("Category", "gamma"), MefV2.Shared] public class Service42 : IService { public string Name => "42"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 43), MefV2.ExportMetadata("Category", "delta"), MefV2.Shared] public class Service43 : IService { public string Name => "43"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 44), MefV2.ExportMetadata("Category", "alpha"), MefV2.Shared] public class Service44 : IService { public string Name => "44"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 45), MefV2.ExportMetadata("Category", "beta"), MefV2.Shared] public class Service45 : IService { public string Name => "45"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 46), MefV2.ExportMetadata("Category", "gamma"), MefV2.Shared] public class Service46 : IService { public string Name => "46"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 47), MefV2.ExportMetadata("Category", "delta"), MefV2.Shared] public class Service47 : IService { public string Name => "47"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 48), MefV2.ExportMetadata("Category", "alpha"), MefV2.Shared] public class Service48 : IService { public string Name => "48"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 49), MefV2.ExportMetadata("Category", "beta"), MefV2.Shared] public class Service49 : IService { public string Name => "49"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 50), MefV2.ExportMetadata("Category", "gamma"), MefV2.Shared] public class Service50 : IService { public string Name => "50"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 51), MefV2.ExportMetadata("Category", "delta"), MefV2.Shared] public class Service51 : IService { public string Name => "51"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 52), MefV2.ExportMetadata("Category", "alpha"), MefV2.Shared] public class Service52 : IService { public string Name => "52"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 53), MefV2.ExportMetadata("Category", "beta"), MefV2.Shared] public class Service53 : IService { public string Name => "53"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 54), MefV2.ExportMetadata("Category", "gamma"), MefV2.Shared] public class Service54 : IService { public string Name => "54"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 55), MefV2.ExportMetadata("Category", "delta"), MefV2.Shared] public class Service55 : IService { public string Name => "55"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 56), MefV2.ExportMetadata("Category", "alpha"), MefV2.Shared] public class Service56 : IService { public string Name => "56"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 57), MefV2.ExportMetadata("Category", "beta"), MefV2.Shared] public class Service57 : IService { public string Name => "57"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 58), MefV2.ExportMetadata("Category", "gamma"), MefV2.Shared] public class Service58 : IService { public string Name => "58"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 59), MefV2.ExportMetadata("Category", "delta"), MefV2.Shared] public class Service59 : IService { public string Name => "59"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 60), MefV2.ExportMetadata("Category", "alpha"), MefV2.Shared] public class Service60 : IService { public string Name => "60"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 61), MefV2.ExportMetadata("Category", "beta"), MefV2.Shared] public class Service61 : IService { public string Name => "61"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 62), MefV2.ExportMetadata("Category", "gamma"), MefV2.Shared] public class Service62 : IService { public string Name => "62"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 63), MefV2.ExportMetadata("Category", "delta"), MefV2.Shared] public class Service63 : IService { public string Name => "63"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 64), MefV2.ExportMetadata("Category", "alpha"), MefV2.Shared] public class Service64 : IService { public string Name => "64"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 65), MefV2.ExportMetadata("Category", "beta"), MefV2.Shared] public class Service65 : IService { public string Name => "65"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 66), MefV2.ExportMetadata("Category", "gamma"), MefV2.Shared] public class Service66 : IService { public string Name => "66"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 67), MefV2.ExportMetadata("Category", "delta"), MefV2.Shared] public class Service67 : IService { public string Name => "67"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 68), MefV2.ExportMetadata("Category", "alpha"), MefV2.Shared] public class Service68 : IService { public string Name => "68"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 69), MefV2.ExportMetadata("Category", "beta"), MefV2.Shared] public class Service69 : IService { public string Name => "69"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 70), MefV2.ExportMetadata("Category", "gamma"), MefV2.Shared] public class Service70 : IService { public string Name => "70"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 71), MefV2.ExportMetadata("Category", "delta"), MefV2.Shared] public class Service71 : IService { public string Name => "71"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 72), MefV2.ExportMetadata("Category", "alpha"), MefV2.Shared] public class Service72 : IService { public string Name => "72"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 73), MefV2.ExportMetadata("Category", "beta"), MefV2.Shared] public class Service73 : IService { public string Name => "73"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 74), MefV2.ExportMetadata("Category", "gamma"), MefV2.Shared] public class Service74 : IService { public string Name => "74"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 75), MefV2.ExportMetadata("Category", "delta"), MefV2.Shared] public class Service75 : IService { public string Name => "75"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 76), MefV2.ExportMetadata("Category", "alpha"), MefV2.Shared] public class Service76 : IService { public string Name => "76"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 77), MefV2.ExportMetadata("Category", "beta"), MefV2.Shared] public class Service77 : IService { public string Name => "77"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 78), MefV2.ExportMetadata("Category", "gamma"), MefV2.Shared] public class Service78 : IService { public string Name => "78"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 79), MefV2.ExportMetadata("Category", "delta"), MefV2.Shared] public class Service79 : IService { public string Name => "79"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 80), MefV2.ExportMetadata("Category", "alpha"), MefV2.Shared] public class Service80 : IService { public string Name => "80"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 81), MefV2.ExportMetadata("Category", "beta"), MefV2.Shared] public class Service81 : IService { public string Name => "81"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 82), MefV2.ExportMetadata("Category", "gamma"), MefV2.Shared] public class Service82 : IService { public string Name => "82"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 83), MefV2.ExportMetadata("Category", "delta"), MefV2.Shared] public class Service83 : IService { public string Name => "83"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 84), MefV2.ExportMetadata("Category", "alpha"), MefV2.Shared] public class Service84 : IService { public string Name => "84"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 85), MefV2.ExportMetadata("Category", "beta"), MefV2.Shared] public class Service85 : IService { public string Name => "85"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 86), MefV2.ExportMetadata("Category", "gamma"), MefV2.Shared] public class Service86 : IService { public string Name => "86"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 87), MefV2.ExportMetadata("Category", "delta"), MefV2.Shared] public class Service87 : IService { public string Name => "87"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 88), MefV2.ExportMetadata("Category", "alpha"), MefV2.Shared] public class Service88 : IService { public string Name => "88"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 89), MefV2.ExportMetadata("Category", "beta"), MefV2.Shared] public class Service89 : IService { public string Name => "89"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 90), MefV2.ExportMetadata("Category", "gamma"), MefV2.Shared] public class Service90 : IService { public string Name => "90"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 91), MefV2.ExportMetadata("Category", "delta"), MefV2.Shared] public class Service91 : IService { public string Name => "91"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 92), MefV2.ExportMetadata("Category", "alpha"), MefV2.Shared] public class Service92 : IService { public string Name => "92"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 93), MefV2.ExportMetadata("Category", "beta"), MefV2.Shared] public class Service93 : IService { public string Name => "93"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 94), MefV2.ExportMetadata("Category", "gamma"), MefV2.Shared] public class Service94 : IService { public string Name => "94"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 95), MefV2.ExportMetadata("Category", "delta"), MefV2.Shared] public class Service95 : IService { public string Name => "95"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 96), MefV2.ExportMetadata("Category", "alpha"), MefV2.Shared] public class Service96 : IService { public string Name => "96"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 97), MefV2.ExportMetadata("Category", "beta"), MefV2.Shared] public class Service97 : IService { public string Name => "97"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 98), MefV2.ExportMetadata("Category", "gamma"), MefV2.Shared] public class Service98 : IService { public string Name => "98"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 99), MefV2.ExportMetadata("Category", "delta"), MefV2.Shared] public class Service99 : IService { public string Name => "99"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 100), MefV2.ExportMetadata("Category", "alpha"), MefV2.Shared] public class Service100 : IService { public string Name => "100"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 101), MefV2.ExportMetadata("Category", "beta"), MefV2.Shared] public class Service101 : IService { public string Name => "101"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 102), MefV2.ExportMetadata("Category", "gamma"), MefV2.Shared] public class Service102 : IService { public string Name => "102"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 103), MefV2.ExportMetadata("Category", "delta"), MefV2.Shared] public class Service103 : IService { public string Name => "103"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 104), MefV2.ExportMetadata("Category", "alpha"), MefV2.Shared] public class Service104 : IService { public string Name => "104"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 105), MefV2.ExportMetadata("Category", "beta"), MefV2.Shared] public class Service105 : IService { public string Name => "105"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 106), MefV2.ExportMetadata("Category", "gamma"), MefV2.Shared] public class Service106 : IService { public string Name => "106"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 107), MefV2.ExportMetadata("Category", "delta"), MefV2.Shared] public class Service107 : IService { public string Name => "107"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 108), MefV2.ExportMetadata("Category", "alpha"), MefV2.Shared] public class Service108 : IService { public string Name => "108"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 109), MefV2.ExportMetadata("Category", "beta"), MefV2.Shared] public class Service109 : IService { public string Name => "109"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 110), MefV2.ExportMetadata("Category", "gamma"), MefV2.Shared] public class Service110 : IService { public string Name => "110"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 111), MefV2.ExportMetadata("Category", "delta"), MefV2.Shared] public class Service111 : IService { public string Name => "111"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 112), MefV2.ExportMetadata("Category", "alpha"), MefV2.Shared] public class Service112 : IService { public string Name => "112"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 113), MefV2.ExportMetadata("Category", "beta"), MefV2.Shared] public class Service113 : IService { public string Name => "113"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 114), MefV2.ExportMetadata("Category", "gamma"), MefV2.Shared] public class Service114 : IService { public string Name => "114"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 115), MefV2.ExportMetadata("Category", "delta"), MefV2.Shared] public class Service115 : IService { public string Name => "115"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 116), MefV2.ExportMetadata("Category", "alpha"), MefV2.Shared] public class Service116 : IService { public string Name => "116"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 117), MefV2.ExportMetadata("Category", "beta"), MefV2.Shared] public class Service117 : IService { public string Name => "117"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 118), MefV2.ExportMetadata("Category", "gamma"), MefV2.Shared] public class Service118 : IService { public string Name => "118"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 119), MefV2.ExportMetadata("Category", "delta"), MefV2.Shared] public class Service119 : IService { public string Name => "119"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 120), MefV2.ExportMetadata("Category", "alpha"), MefV2.Shared] public class Service120 : IService { public string Name => "120"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 121), MefV2.ExportMetadata("Category", "beta"), MefV2.Shared] public class Service121 : IService { public string Name => "121"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 122), MefV2.ExportMetadata("Category", "gamma"), MefV2.Shared] public class Service122 : IService { public string Name => "122"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 123), MefV2.ExportMetadata("Category", "delta"), MefV2.Shared] public class Service123 : IService { public string Name => "123"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 124), MefV2.ExportMetadata("Category", "alpha"), MefV2.Shared] public class Service124 : IService { public string Name => "124"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 125), MefV2.ExportMetadata("Category", "beta"), MefV2.Shared] public class Service125 : IService { public string Name => "125"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 126), MefV2.ExportMetadata("Category", "gamma"), MefV2.Shared] public class Service126 : IService { public string Name => "126"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 127), MefV2.ExportMetadata("Category", "delta"), MefV2.Shared] public class Service127 : IService { public string Name => "127"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 128), MefV2.ExportMetadata("Category", "alpha"), MefV2.Shared] public class Service128 : IService { public string Name => "128"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 129), MefV2.ExportMetadata("Category", "beta"), MefV2.Shared] public class Service129 : IService { public string Name => "129"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 130), MefV2.ExportMetadata("Category", "gamma"), MefV2.Shared] public class Service130 : IService { public string Name => "130"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 131), MefV2.ExportMetadata("Category", "delta"), MefV2.Shared] public class Service131 : IService { public string Name => "131"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 132), MefV2.ExportMetadata("Category", "alpha"), MefV2.Shared] public class Service132 : IService { public string Name => "132"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 133), MefV2.ExportMetadata("Category", "beta"), MefV2.Shared] public class Service133 : IService { public string Name => "133"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 134), MefV2.ExportMetadata("Category", "gamma"), MefV2.Shared] public class Service134 : IService { public string Name => "134"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 135), MefV2.ExportMetadata("Category", "delta"), MefV2.Shared] public class Service135 : IService { public string Name => "135"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 136), MefV2.ExportMetadata("Category", "alpha"), MefV2.Shared] public class Service136 : IService { public string Name => "136"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 137), MefV2.ExportMetadata("Category", "beta"), MefV2.Shared] public class Service137 : IService { public string Name => "137"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 138), MefV2.ExportMetadata("Category", "gamma"), MefV2.Shared] public class Service138 : IService { public string Name => "138"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 139), MefV2.ExportMetadata("Category", "delta"), MefV2.Shared] public class Service139 : IService { public string Name => "139"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 140), MefV2.ExportMetadata("Category", "alpha"), MefV2.Shared] public class Service140 : IService { public string Name => "140"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 141), MefV2.ExportMetadata("Category", "beta"), MefV2.Shared] public class Service141 : IService { public string Name => "141"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 142), MefV2.ExportMetadata("Category", "gamma"), MefV2.Shared] public class Service142 : IService { public string Name => "142"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 143), MefV2.ExportMetadata("Category", "delta"), MefV2.Shared] public class Service143 : IService { public string Name => "143"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 144), MefV2.ExportMetadata("Category", "alpha"), MefV2.Shared] public class Service144 : IService { public string Name => "144"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 145), MefV2.ExportMetadata("Category", "beta"), MefV2.Shared] public class Service145 : IService { public string Name => "145"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 146), MefV2.ExportMetadata("Category", "gamma"), MefV2.Shared] public class Service146 : IService { public string Name => "146"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 147), MefV2.ExportMetadata("Category", "delta"), MefV2.Shared] public class Service147 : IService { public string Name => "147"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 148), MefV2.ExportMetadata("Category", "alpha"), MefV2.Shared] public class Service148 : IService { public string Name => "148"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 149), MefV2.ExportMetadata("Category", "beta"), MefV2.Shared] public class Service149 : IService { public string Name => "149"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 150), MefV2.ExportMetadata("Category", "gamma"), MefV2.Shared] public class Service150 : IService { public string Name => "150"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 151), MefV2.ExportMetadata("Category", "delta"), MefV2.Shared] public class Service151 : IService { public string Name => "151"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 152), MefV2.ExportMetadata("Category", "alpha"), MefV2.Shared] public class Service152 : IService { public string Name => "152"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 153), MefV2.ExportMetadata("Category", "beta"), MefV2.Shared] public class Service153 : IService { public string Name => "153"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 154), MefV2.ExportMetadata("Category", "gamma"), MefV2.Shared] public class Service154 : IService { public string Name => "154"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 155), MefV2.ExportMetadata("Category", "delta"), MefV2.Shared] public class Service155 : IService { public string Name => "155"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 156), MefV2.ExportMetadata("Category", "alpha"), MefV2.Shared] public class Service156 : IService { public string Name => "156"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 157), MefV2.ExportMetadata("Category", "beta"), MefV2.Shared] public class Service157 : IService { public string Name => "157"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 158), MefV2.ExportMetadata("Category", "gamma"), MefV2.Shared] public class Service158 : IService { public string Name => "158"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 159), MefV2.ExportMetadata("Category", "delta"), MefV2.Shared] public class Service159 : IService { public string Name => "159"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 160), MefV2.ExportMetadata("Category", "alpha"), MefV2.Shared] public class Service160 : IService { public string Name => "160"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 161), MefV2.ExportMetadata("Category", "beta"), MefV2.Shared] public class Service161 : IService { public string Name => "161"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 162), MefV2.ExportMetadata("Category", "gamma"), MefV2.Shared] public class Service162 : IService { public string Name => "162"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 163), MefV2.ExportMetadata("Category", "delta"), MefV2.Shared] public class Service163 : IService { public string Name => "163"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 164), MefV2.ExportMetadata("Category", "alpha"), MefV2.Shared] public class Service164 : IService { public string Name => "164"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 165), MefV2.ExportMetadata("Category", "beta"), MefV2.Shared] public class Service165 : IService { public string Name => "165"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 166), MefV2.ExportMetadata("Category", "gamma"), MefV2.Shared] public class Service166 : IService { public string Name => "166"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 167), MefV2.ExportMetadata("Category", "delta"), MefV2.Shared] public class Service167 : IService { public string Name => "167"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 168), MefV2.ExportMetadata("Category", "alpha"), MefV2.Shared] public class Service168 : IService { public string Name => "168"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 169), MefV2.ExportMetadata("Category", "beta"), MefV2.Shared] public class Service169 : IService { public string Name => "169"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 170), MefV2.ExportMetadata("Category", "gamma"), MefV2.Shared] public class Service170 : IService { public string Name => "170"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 171), MefV2.ExportMetadata("Category", "delta"), MefV2.Shared] public class Service171 : IService { public string Name => "171"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 172), MefV2.ExportMetadata("Category", "alpha"), MefV2.Shared] public class Service172 : IService { public string Name => "172"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 173), MefV2.ExportMetadata("Category", "beta"), MefV2.Shared] public class Service173 : IService { public string Name => "173"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 174), MefV2.ExportMetadata("Category", "gamma"), MefV2.Shared] public class Service174 : IService { public string Name => "174"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 175), MefV2.ExportMetadata("Category", "delta"), MefV2.Shared] public class Service175 : IService { public string Name => "175"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 176), MefV2.ExportMetadata("Category", "alpha"), MefV2.Shared] public class Service176 : IService { public string Name => "176"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 177), MefV2.ExportMetadata("Category", "beta"), MefV2.Shared] public class Service177 : IService { public string Name => "177"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 178), MefV2.ExportMetadata("Category", "gamma"), MefV2.Shared] public class Service178 : IService { public string Name => "178"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 179), MefV2.ExportMetadata("Category", "delta"), MefV2.Shared] public class Service179 : IService { public string Name => "179"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 180), MefV2.ExportMetadata("Category", "alpha"), MefV2.Shared] public class Service180 : IService { public string Name => "180"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 181), MefV2.ExportMetadata("Category", "beta"), MefV2.Shared] public class Service181 : IService { public string Name => "181"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 182), MefV2.ExportMetadata("Category", "gamma"), MefV2.Shared] public class Service182 : IService { public string Name => "182"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 183), MefV2.ExportMetadata("Category", "delta"), MefV2.Shared] public class Service183 : IService { public string Name => "183"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 184), MefV2.ExportMetadata("Category", "alpha"), MefV2.Shared] public class Service184 : IService { public string Name => "184"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 185), MefV2.ExportMetadata("Category", "beta"), MefV2.Shared] public class Service185 : IService { public string Name => "185"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 186), MefV2.ExportMetadata("Category", "gamma"), MefV2.Shared] public class Service186 : IService { public string Name => "186"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 187), MefV2.ExportMetadata("Category", "delta"), MefV2.Shared] public class Service187 : IService { public string Name => "187"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 188), MefV2.ExportMetadata("Category", "alpha"), MefV2.Shared] public class Service188 : IService { public string Name => "188"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 189), MefV2.ExportMetadata("Category", "beta"), MefV2.Shared] public class Service189 : IService { public string Name => "189"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 190), MefV2.ExportMetadata("Category", "gamma"), MefV2.Shared] public class Service190 : IService { public string Name => "190"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 191), MefV2.ExportMetadata("Category", "delta"), MefV2.Shared] public class Service191 : IService { public string Name => "191"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 192), MefV2.ExportMetadata("Category", "alpha"), MefV2.Shared] public class Service192 : IService { public string Name => "192"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 193), MefV2.ExportMetadata("Category", "beta"), MefV2.Shared] public class Service193 : IService { public string Name => "193"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 194), MefV2.ExportMetadata("Category", "gamma"), MefV2.Shared] public class Service194 : IService { public string Name => "194"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 195), MefV2.ExportMetadata("Category", "delta"), MefV2.Shared] public class Service195 : IService { public string Name => "195"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 196), MefV2.ExportMetadata("Category", "alpha"), MefV2.Shared] public class Service196 : IService { public string Name => "196"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 197), MefV2.ExportMetadata("Category", "beta"), MefV2.Shared] public class Service197 : IService { public string Name => "197"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 198), MefV2.ExportMetadata("Category", "gamma"), MefV2.Shared] public class Service198 : IService { public string Name => "198"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 199), MefV2.ExportMetadata("Category", "delta"), MefV2.Shared] public class Service199 : IService { public string Name => "199"; }
-    [MefV2.Export(typeof(IService)), MefV2.ExportMetadata("Order", 200), MefV2.ExportMetadata("Category", "alpha"), MefV2.Shared] public class Service200 : IService { public string Name => "200"; }
-
-    public abstract class ProcessorBase : IProcessor
-    {
-        protected ProcessorBase(ILogger logger, ICache cache) { this.Logger = logger; this.Cache = cache; }
-        protected ILogger Logger { get; }
-        protected ICache Cache { get; }
-        [MefV2.ImportMany] public IEnumerable<System.Lazy<IService, IServiceMetadata>> Services { get; set; }
-        public string Process(string input) => input;
-    }
-
-    [MefV2.Export(typeof(IProcessor)), MefV2.Shared] public class Processor1 : ProcessorBase { [MefV2.ImportingConstructor] public Processor1(ILogger l, ICache c) : base(l, c) { } }
-    [MefV2.Export(typeof(IProcessor)), MefV2.Shared] public class Processor2 : ProcessorBase { [MefV2.ImportingConstructor] public Processor2(ILogger l, ICache c) : base(l, c) { } }
-    [MefV2.Export(typeof(IProcessor)), MefV2.Shared] public class Processor3 : ProcessorBase { [MefV2.ImportingConstructor] public Processor3(ILogger l, ICache c) : base(l, c) { } }
-    [MefV2.Export(typeof(IProcessor)), MefV2.Shared] public class Processor4 : ProcessorBase { [MefV2.ImportingConstructor] public Processor4(ILogger l, ICache c) : base(l, c) { } }
-    [MefV2.Export(typeof(IProcessor)), MefV2.Shared] public class Processor5 : ProcessorBase { [MefV2.ImportingConstructor] public Processor5(ILogger l, ICache c) : base(l, c) { } }
-    [MefV2.Export(typeof(IProcessor)), MefV2.Shared] public class Processor6 : ProcessorBase { [MefV2.ImportingConstructor] public Processor6(ILogger l, ICache c) : base(l, c) { } }
-    [MefV2.Export(typeof(IProcessor)), MefV2.Shared] public class Processor7 : ProcessorBase { [MefV2.ImportingConstructor] public Processor7(ILogger l, ICache c) : base(l, c) { } }
-    [MefV2.Export(typeof(IProcessor)), MefV2.Shared] public class Processor8 : ProcessorBase { [MefV2.ImportingConstructor] public Processor8(ILogger l, ICache c) : base(l, c) { } }
-    [MefV2.Export(typeof(IProcessor)), MefV2.Shared] public class Processor9 : ProcessorBase { [MefV2.ImportingConstructor] public Processor9(ILogger l, ICache c) : base(l, c) { } }
-    [MefV2.Export(typeof(IProcessor)), MefV2.Shared] public class Processor10 : ProcessorBase { [MefV2.ImportingConstructor] public Processor10(ILogger l, ICache c) : base(l, c) { } }
-    [MefV2.Export(typeof(IProcessor)), MefV2.Shared] public class Processor11 : ProcessorBase { [MefV2.ImportingConstructor] public Processor11(ILogger l, ICache c) : base(l, c) { } }
-    [MefV2.Export(typeof(IProcessor)), MefV2.Shared] public class Processor12 : ProcessorBase { [MefV2.ImportingConstructor] public Processor12(ILogger l, ICache c) : base(l, c) { } }
-    [MefV2.Export(typeof(IProcessor)), MefV2.Shared] public class Processor13 : ProcessorBase { [MefV2.ImportingConstructor] public Processor13(ILogger l, ICache c) : base(l, c) { } }
-    [MefV2.Export(typeof(IProcessor)), MefV2.Shared] public class Processor14 : ProcessorBase { [MefV2.ImportingConstructor] public Processor14(ILogger l, ICache c) : base(l, c) { } }
-    [MefV2.Export(typeof(IProcessor)), MefV2.Shared] public class Processor15 : ProcessorBase { [MefV2.ImportingConstructor] public Processor15(ILogger l, ICache c) : base(l, c) { } }
-    [MefV2.Export(typeof(IProcessor)), MefV2.Shared] public class Processor16 : ProcessorBase { [MefV2.ImportingConstructor] public Processor16(ILogger l, ICache c) : base(l, c) { } }
-    [MefV2.Export(typeof(IProcessor)), MefV2.Shared] public class Processor17 : ProcessorBase { [MefV2.ImportingConstructor] public Processor17(ILogger l, ICache c) : base(l, c) { } }
-    [MefV2.Export(typeof(IProcessor)), MefV2.Shared] public class Processor18 : ProcessorBase { [MefV2.ImportingConstructor] public Processor18(ILogger l, ICache c) : base(l, c) { } }
-    [MefV2.Export(typeof(IProcessor)), MefV2.Shared] public class Processor19 : ProcessorBase { [MefV2.ImportingConstructor] public Processor19(ILogger l, ICache c) : base(l, c) { } }
-    [MefV2.Export(typeof(IProcessor)), MefV2.Shared] public class Processor20 : ProcessorBase { [MefV2.ImportingConstructor] public Processor20(ILogger l, ICache c) : base(l, c) { } }
-    [MefV2.Export(typeof(IProcessor)), MefV2.Shared] public class Processor21 : ProcessorBase { [MefV2.ImportingConstructor] public Processor21(ILogger l, ICache c) : base(l, c) { } }
-    [MefV2.Export(typeof(IProcessor)), MefV2.Shared] public class Processor22 : ProcessorBase { [MefV2.ImportingConstructor] public Processor22(ILogger l, ICache c) : base(l, c) { } }
-    [MefV2.Export(typeof(IProcessor)), MefV2.Shared] public class Processor23 : ProcessorBase { [MefV2.ImportingConstructor] public Processor23(ILogger l, ICache c) : base(l, c) { } }
-    [MefV2.Export(typeof(IProcessor)), MefV2.Shared] public class Processor24 : ProcessorBase { [MefV2.ImportingConstructor] public Processor24(ILogger l, ICache c) : base(l, c) { } }
-    [MefV2.Export(typeof(IProcessor)), MefV2.Shared] public class Processor25 : ProcessorBase { [MefV2.ImportingConstructor] public Processor25(ILogger l, ICache c) : base(l, c) { } }
-    [MefV2.Export(typeof(IProcessor)), MefV2.Shared] public class Processor26 : ProcessorBase { [MefV2.ImportingConstructor] public Processor26(ILogger l, ICache c) : base(l, c) { } }
-    [MefV2.Export(typeof(IProcessor)), MefV2.Shared] public class Processor27 : ProcessorBase { [MefV2.ImportingConstructor] public Processor27(ILogger l, ICache c) : base(l, c) { } }
-    [MefV2.Export(typeof(IProcessor)), MefV2.Shared] public class Processor28 : ProcessorBase { [MefV2.ImportingConstructor] public Processor28(ILogger l, ICache c) : base(l, c) { } }
-    [MefV2.Export(typeof(IProcessor)), MefV2.Shared] public class Processor29 : ProcessorBase { [MefV2.ImportingConstructor] public Processor29(ILogger l, ICache c) : base(l, c) { } }
-    [MefV2.Export(typeof(IProcessor)), MefV2.Shared] public class Processor30 : ProcessorBase { [MefV2.ImportingConstructor] public Processor30(ILogger l, ICache c) : base(l, c) { } }
-    [MefV2.Export(typeof(IProcessor)), MefV2.Shared] public class Processor31 : ProcessorBase { [MefV2.ImportingConstructor] public Processor31(ILogger l, ICache c) : base(l, c) { } }
-    [MefV2.Export(typeof(IProcessor)), MefV2.Shared] public class Processor32 : ProcessorBase { [MefV2.ImportingConstructor] public Processor32(ILogger l, ICache c) : base(l, c) { } }
-    [MefV2.Export(typeof(IProcessor)), MefV2.Shared] public class Processor33 : ProcessorBase { [MefV2.ImportingConstructor] public Processor33(ILogger l, ICache c) : base(l, c) { } }
-    [MefV2.Export(typeof(IProcessor)), MefV2.Shared] public class Processor34 : ProcessorBase { [MefV2.ImportingConstructor] public Processor34(ILogger l, ICache c) : base(l, c) { } }
-    [MefV2.Export(typeof(IProcessor)), MefV2.Shared] public class Processor35 : ProcessorBase { [MefV2.ImportingConstructor] public Processor35(ILogger l, ICache c) : base(l, c) { } }
-    [MefV2.Export(typeof(IProcessor)), MefV2.Shared] public class Processor36 : ProcessorBase { [MefV2.ImportingConstructor] public Processor36(ILogger l, ICache c) : base(l, c) { } }
-    [MefV2.Export(typeof(IProcessor)), MefV2.Shared] public class Processor37 : ProcessorBase { [MefV2.ImportingConstructor] public Processor37(ILogger l, ICache c) : base(l, c) { } }
-    [MefV2.Export(typeof(IProcessor)), MefV2.Shared] public class Processor38 : ProcessorBase { [MefV2.ImportingConstructor] public Processor38(ILogger l, ICache c) : base(l, c) { } }
-    [MefV2.Export(typeof(IProcessor)), MefV2.Shared] public class Processor39 : ProcessorBase { [MefV2.ImportingConstructor] public Processor39(ILogger l, ICache c) : base(l, c) { } }
-    [MefV2.Export(typeof(IProcessor)), MefV2.Shared] public class Processor40 : ProcessorBase { [MefV2.ImportingConstructor] public Processor40(ILogger l, ICache c) : base(l, c) { } }
-    [MefV2.Export(typeof(IProcessor)), MefV2.Shared] public class Processor41 : ProcessorBase { [MefV2.ImportingConstructor] public Processor41(ILogger l, ICache c) : base(l, c) { } }
-    [MefV2.Export(typeof(IProcessor)), MefV2.Shared] public class Processor42 : ProcessorBase { [MefV2.ImportingConstructor] public Processor42(ILogger l, ICache c) : base(l, c) { } }
-    [MefV2.Export(typeof(IProcessor)), MefV2.Shared] public class Processor43 : ProcessorBase { [MefV2.ImportingConstructor] public Processor43(ILogger l, ICache c) : base(l, c) { } }
-    [MefV2.Export(typeof(IProcessor)), MefV2.Shared] public class Processor44 : ProcessorBase { [MefV2.ImportingConstructor] public Processor44(ILogger l, ICache c) : base(l, c) { } }
-    [MefV2.Export(typeof(IProcessor)), MefV2.Shared] public class Processor45 : ProcessorBase { [MefV2.ImportingConstructor] public Processor45(ILogger l, ICache c) : base(l, c) { } }
-    [MefV2.Export(typeof(IProcessor)), MefV2.Shared] public class Processor46 : ProcessorBase { [MefV2.ImportingConstructor] public Processor46(ILogger l, ICache c) : base(l, c) { } }
-    [MefV2.Export(typeof(IProcessor)), MefV2.Shared] public class Processor47 : ProcessorBase { [MefV2.ImportingConstructor] public Processor47(ILogger l, ICache c) : base(l, c) { } }
-    [MefV2.Export(typeof(IProcessor)), MefV2.Shared] public class Processor48 : ProcessorBase { [MefV2.ImportingConstructor] public Processor48(ILogger l, ICache c) : base(l, c) { } }
-    [MefV2.Export(typeof(IProcessor)), MefV2.Shared] public class Processor49 : ProcessorBase { [MefV2.ImportingConstructor] public Processor49(ILogger l, ICache c) : base(l, c) { } }
-    [MefV2.Export(typeof(IProcessor)), MefV2.Shared] public class Processor50 : ProcessorBase { [MefV2.ImportingConstructor] public Processor50(ILogger l, ICache c) : base(l, c) { } }
-    [MefV2.Export(typeof(IProcessor)), MefV2.Shared] public class Processor51 : ProcessorBase { [MefV2.ImportingConstructor] public Processor51(ILogger l, ICache c) : base(l, c) { } }
-    [MefV2.Export(typeof(IProcessor)), MefV2.Shared] public class Processor52 : ProcessorBase { [MefV2.ImportingConstructor] public Processor52(ILogger l, ICache c) : base(l, c) { } }
-    [MefV2.Export(typeof(IProcessor)), MefV2.Shared] public class Processor53 : ProcessorBase { [MefV2.ImportingConstructor] public Processor53(ILogger l, ICache c) : base(l, c) { } }
-    [MefV2.Export(typeof(IProcessor)), MefV2.Shared] public class Processor54 : ProcessorBase { [MefV2.ImportingConstructor] public Processor54(ILogger l, ICache c) : base(l, c) { } }
-    [MefV2.Export(typeof(IProcessor)), MefV2.Shared] public class Processor55 : ProcessorBase { [MefV2.ImportingConstructor] public Processor55(ILogger l, ICache c) : base(l, c) { } }
-    [MefV2.Export(typeof(IProcessor)), MefV2.Shared] public class Processor56 : ProcessorBase { [MefV2.ImportingConstructor] public Processor56(ILogger l, ICache c) : base(l, c) { } }
-    [MefV2.Export(typeof(IProcessor)), MefV2.Shared] public class Processor57 : ProcessorBase { [MefV2.ImportingConstructor] public Processor57(ILogger l, ICache c) : base(l, c) { } }
-    [MefV2.Export(typeof(IProcessor)), MefV2.Shared] public class Processor58 : ProcessorBase { [MefV2.ImportingConstructor] public Processor58(ILogger l, ICache c) : base(l, c) { } }
-    [MefV2.Export(typeof(IProcessor)), MefV2.Shared] public class Processor59 : ProcessorBase { [MefV2.ImportingConstructor] public Processor59(ILogger l, ICache c) : base(l, c) { } }
-    [MefV2.Export(typeof(IProcessor)), MefV2.Shared] public class Processor60 : ProcessorBase { [MefV2.ImportingConstructor] public Processor60(ILogger l, ICache c) : base(l, c) { } }
-    [MefV2.Export(typeof(IProcessor)), MefV2.Shared] public class Processor61 : ProcessorBase { [MefV2.ImportingConstructor] public Processor61(ILogger l, ICache c) : base(l, c) { } }
-    [MefV2.Export(typeof(IProcessor)), MefV2.Shared] public class Processor62 : ProcessorBase { [MefV2.ImportingConstructor] public Processor62(ILogger l, ICache c) : base(l, c) { } }
-    [MefV2.Export(typeof(IProcessor)), MefV2.Shared] public class Processor63 : ProcessorBase { [MefV2.ImportingConstructor] public Processor63(ILogger l, ICache c) : base(l, c) { } }
-    [MefV2.Export(typeof(IProcessor)), MefV2.Shared] public class Processor64 : ProcessorBase { [MefV2.ImportingConstructor] public Processor64(ILogger l, ICache c) : base(l, c) { } }
-    [MefV2.Export(typeof(IProcessor)), MefV2.Shared] public class Processor65 : ProcessorBase { [MefV2.ImportingConstructor] public Processor65(ILogger l, ICache c) : base(l, c) { } }
-    [MefV2.Export(typeof(IProcessor)), MefV2.Shared] public class Processor66 : ProcessorBase { [MefV2.ImportingConstructor] public Processor66(ILogger l, ICache c) : base(l, c) { } }
-    [MefV2.Export(typeof(IProcessor)), MefV2.Shared] public class Processor67 : ProcessorBase { [MefV2.ImportingConstructor] public Processor67(ILogger l, ICache c) : base(l, c) { } }
-    [MefV2.Export(typeof(IProcessor)), MefV2.Shared] public class Processor68 : ProcessorBase { [MefV2.ImportingConstructor] public Processor68(ILogger l, ICache c) : base(l, c) { } }
-    [MefV2.Export(typeof(IProcessor)), MefV2.Shared] public class Processor69 : ProcessorBase { [MefV2.ImportingConstructor] public Processor69(ILogger l, ICache c) : base(l, c) { } }
-    [MefV2.Export(typeof(IProcessor)), MefV2.Shared] public class Processor70 : ProcessorBase { [MefV2.ImportingConstructor] public Processor70(ILogger l, ICache c) : base(l, c) { } }
-    [MefV2.Export(typeof(IProcessor)), MefV2.Shared] public class Processor71 : ProcessorBase { [MefV2.ImportingConstructor] public Processor71(ILogger l, ICache c) : base(l, c) { } }
-    [MefV2.Export(typeof(IProcessor)), MefV2.Shared] public class Processor72 : ProcessorBase { [MefV2.ImportingConstructor] public Processor72(ILogger l, ICache c) : base(l, c) { } }
-    [MefV2.Export(typeof(IProcessor)), MefV2.Shared] public class Processor73 : ProcessorBase { [MefV2.ImportingConstructor] public Processor73(ILogger l, ICache c) : base(l, c) { } }
-    [MefV2.Export(typeof(IProcessor)), MefV2.Shared] public class Processor74 : ProcessorBase { [MefV2.ImportingConstructor] public Processor74(ILogger l, ICache c) : base(l, c) { } }
-    [MefV2.Export(typeof(IProcessor)), MefV2.Shared] public class Processor75 : ProcessorBase { [MefV2.ImportingConstructor] public Processor75(ILogger l, ICache c) : base(l, c) { } }
-    [MefV2.Export(typeof(IProcessor)), MefV2.Shared] public class Processor76 : ProcessorBase { [MefV2.ImportingConstructor] public Processor76(ILogger l, ICache c) : base(l, c) { } }
-    [MefV2.Export(typeof(IProcessor)), MefV2.Shared] public class Processor77 : ProcessorBase { [MefV2.ImportingConstructor] public Processor77(ILogger l, ICache c) : base(l, c) { } }
-    [MefV2.Export(typeof(IProcessor)), MefV2.Shared] public class Processor78 : ProcessorBase { [MefV2.ImportingConstructor] public Processor78(ILogger l, ICache c) : base(l, c) { } }
-    [MefV2.Export(typeof(IProcessor)), MefV2.Shared] public class Processor79 : ProcessorBase { [MefV2.ImportingConstructor] public Processor79(ILogger l, ICache c) : base(l, c) { } }
-    [MefV2.Export(typeof(IProcessor)), MefV2.Shared] public class Processor80 : ProcessorBase { [MefV2.ImportingConstructor] public Processor80(ILogger l, ICache c) : base(l, c) { } }
-
-    [MefV2.Export] public class Transient1 { [MefV2.ImportingConstructor] public Transient1(ILogger l, ICache c) { } }
-    [MefV2.Export] public class Transient2 { [MefV2.ImportingConstructor] public Transient2(ILogger l, ICache c) { } }
-    [MefV2.Export] public class Transient3 { [MefV2.ImportingConstructor] public Transient3(ILogger l, ICache c) { } }
-    [MefV2.Export] public class Transient4 { [MefV2.ImportingConstructor] public Transient4(ILogger l, ICache c) { } }
-    [MefV2.Export] public class Transient5 { [MefV2.ImportingConstructor] public Transient5(ILogger l, ICache c) { } }
-    [MefV2.Export] public class Transient6 { [MefV2.ImportingConstructor] public Transient6(ILogger l, ICache c) { } }
-    [MefV2.Export] public class Transient7 { [MefV2.ImportingConstructor] public Transient7(ILogger l, ICache c) { } }
-    [MefV2.Export] public class Transient8 { [MefV2.ImportingConstructor] public Transient8(ILogger l, ICache c) { } }
-    [MefV2.Export] public class Transient9 { [MefV2.ImportingConstructor] public Transient9(ILogger l, ICache c) { } }
-    [MefV2.Export] public class Transient10 { [MefV2.ImportingConstructor] public Transient10(ILogger l, ICache c) { } }
-    [MefV2.Export] public class Transient11 { [MefV2.ImportingConstructor] public Transient11(ILogger l, ICache c) { } }
-    [MefV2.Export] public class Transient12 { [MefV2.ImportingConstructor] public Transient12(ILogger l, ICache c) { } }
-    [MefV2.Export] public class Transient13 { [MefV2.ImportingConstructor] public Transient13(ILogger l, ICache c) { } }
-    [MefV2.Export] public class Transient14 { [MefV2.ImportingConstructor] public Transient14(ILogger l, ICache c) { } }
-    [MefV2.Export] public class Transient15 { [MefV2.ImportingConstructor] public Transient15(ILogger l, ICache c) { } }
-    [MefV2.Export] public class Transient16 { [MefV2.ImportingConstructor] public Transient16(ILogger l, ICache c) { } }
-    [MefV2.Export] public class Transient17 { [MefV2.ImportingConstructor] public Transient17(ILogger l, ICache c) { } }
-    [MefV2.Export] public class Transient18 { [MefV2.ImportingConstructor] public Transient18(ILogger l, ICache c) { } }
-    [MefV2.Export] public class Transient19 { [MefV2.ImportingConstructor] public Transient19(ILogger l, ICache c) { } }
-    [MefV2.Export] public class Transient20 { [MefV2.ImportingConstructor] public Transient20(ILogger l, ICache c) { } }
-
-    public interface IServiceV1 { }
-    [MefV1.Export(typeof(IServiceV1))] public class ServiceV1_1 : IServiceV1 { }
-    [MefV1.Export(typeof(IServiceV1))] public class ServiceV1_2 : IServiceV1 { }
-    [MefV1.Export(typeof(IServiceV1))] public class ServiceV1_3 : IServiceV1 { }
-    [MefV1.Export(typeof(IServiceV1))] public class ServiceV1_4 : IServiceV1 { }
-    [MefV1.Export(typeof(IServiceV1))] public class ServiceV1_5 : IServiceV1 { }
-    [MefV1.Export(typeof(IServiceV1))] public class ServiceV1_6 : IServiceV1 { }
-    [MefV1.Export(typeof(IServiceV1))] public class ServiceV1_7 : IServiceV1 { }
-    [MefV1.Export(typeof(IServiceV1))] public class ServiceV1_8 : IServiceV1 { }
-    [MefV1.Export(typeof(IServiceV1))] public class ServiceV1_9 : IServiceV1 { }
-    [MefV1.Export(typeof(IServiceV1))] public class ServiceV1_10 : IServiceV1 { }
-    [MefV1.Export] public class ConsumerV1 { [MefV1.ImportMany] public IEnumerable<IServiceV1> Services { get; set; } }
-}
