--- perf/Microsoft.VisualStudio.Composition.Benchmarks/Program.cs
@@ -1,13 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT license. See LICENSE file in the project root for full license information.
-
-namespace Microsoft.VisualStudio.Composition.Benchmarks
-{
-    using BenchmarkDotNet.Running;
-
-    public static class Program
-    {
-        public static void Main(string[] args) =>
-            BenchmarkSwitcher.FromAssembly(typeof(Program).Assembly).Run(args);
-    }
-}
